Moved compressed image-specific checksum tests to compressed sub-package
